cleanup types

+++ src/include/index/flat_pq_index.h
@@ -79,9 +79,9 @@ class flat_pq_index {
   // metadata
   uint64_t dimensions_{0};
   uint64_t num_subspaces_{0};
-  size_t sub_dimensions_{0};
-  size_t bits_per_subspace_{8};
-  uint64_t num_clusters_{256};
+  uint64_t sub_dimensions_{0};
+  uint32_t bits_per_subspace_{8};
+  uint32_t num_clusters_{256};
   float tol_ = 0.001;
   size_t max_iter_ = 16;
   size_t num_threads_ = std::thread::hardware_concurrency();
@@ -116,8 +116,8 @@ class flat_pq_index {
   flat_pq_index(
       uint64_t dimensions,
       uint64_t num_subspaces,
-      size_t bits_per_subspace = 8,
-      uint64_t num_clusters = 256)
+      uint32_t bits_per_subspace = 8,
+      uint32_t num_clusters = 256)
       : dimensions_(dimensions)
       , num_subspaces_(num_subspaces)
       , bits_per_subspace_(bits_per_subspace)
