+++ src/include/test/unit_ivf_pq_group.cc
@@ -74,7 +74,7 @@ TEST_CASE("write constructor - create and open", "[ivf_pq_group]") {
   }
 
   uint64_t dimensions = 92134;
-  uint64_t num_clusters = 34239;
+  uint32_t num_clusters = 34239;
   uint64_t num_subspaces = 3343;
 
   ivf_pq_group x = ivf_pq_group<dummy_index>(
@@ -192,7 +192,7 @@ TEST_CASE("group metadata - bases, ingestions, partitions", "[ivf_pq_group]") {
   size_t expected_partitions = 42;
   size_t expected_temp_size = 314159;
   uint64_t dimensions = 128;
-  uint64_t num_clusters = 110;
+  uint32_t num_clusters = 110;
   uint64_t num_subspaces = 120;
 }
 
@@ -494,7 +494,7 @@ TEST_CASE("storage version", "[ivf_pq_group]") {
   size_t expected_partitions = 200;
   size_t expected_temp_size = 11;
   uint64_t dimensions = 19238;
-  uint64_t num_clusters = 110;
+  uint32_t num_clusters = 110;
   uint64_t num_subspaces = 120;
   size_t offset = 2345;
 
@@ -574,7 +574,7 @@ TEST_CASE("mismatched storage version", "[ivf_pq_group]") {
   }
 
   uint64_t dimensions = 4;
-  uint64_t num_clusters = 4;
+  uint32_t num_clusters = 4;
   uint64_t num_subspaces = 1;
 
   ivf_pq_group x = ivf_pq_group<dummy_index>(
@@ -609,7 +609,7 @@ TEST_CASE("clear history", "[ivf_pq_group]") {
   }
 
   uint64_t dimensions = 19238;
-  uint64_t num_clusters = 110;
+  uint32_t num_clusters = 110;
   uint64_t num_subspaces = 120;
 
   ivf_pq_group x = ivf_pq_group<dummy_index>(
